requirements.txt
fastapi==0.95.2
uvicorn==0.23.0
torch==1.12.1
numpy==1.23.1
pillow
opencv-python==4.11.0.86
einops==0.3.0
pytorch-lightning==1.5.0
pydantic==2.10.6